+++ packages/moonstone/internal/Picker/Picker.js
@@ -49,7 +49,6 @@ const selectDecIcon = selectIcon('decrementIcon', 'arrowlargedown', 'arrowlargel
 
 // Set-up event forwarding
 const forwardBlur = forward('onBlur'),
-	forwardClick = forward('onClick'),
 	forwardFocus = forward('onFocus'),
 	forwardKeyDown = forward('onKeyDown'),
 	forwardKeyUp = forward('onKeyUp'),
@@ -376,6 +375,9 @@ const PickerBase = class extends React.Component {
 			validateStepped(props.value, props.min, props.step, PickerBase.displayName);
 			validateStepped(props.max, props.min, props.step, PickerBase.displayName, '"max"');
 		}
+
+		// Pressed state for this.handleUp
+		this.pickerButtonPressed = false;
 	}
 
 	componentDidMount () {
@@ -458,20 +460,14 @@ const PickerBase = class extends React.Component {
 		this.reverseTransition = !(dir > 0);
 	}
 
-	handleDecClick = (ev) => {
-		if (ev) {
-			forwardClick(ev, this.props);
-		}
+	handleDecrement = () => {
 		if (!this.hasReachedBound(-this.props.step)) {
 			this.updateValue(-1);
 			this.handleDown(-1);
 		}
 	}
 
-	handleIncClick = (ev) => {
-		if (ev) {
-			forwardClick(ev, this.props);
-		}
+	handleIncrement = () => {
 		if (!this.hasReachedBound(this.props.step)) {
 			this.updateValue(1);
 			this.handleDown(1);
@@ -485,33 +481,30 @@ const PickerBase = class extends React.Component {
 		}
 	}
 
-	emulateMouseUp = new Job((ev) => {
-		const {onMouseUp} = this.props;
-		if (onMouseUp) {
-			onMouseUp(ev);
-		}
-	}, 175)
+	emulateMouseUp = new Job((ev) => forwardMouseUp(ev, this.props), 175)
 
 	handleUp = (ev) => {
-		const {joined} = this.props;
-		forwardMouseUp(ev, this.props);
-		if (joined) {
-			this.emulateMouseUp.start();
+		// Attach handleUp to the container and forward mouseUp if one of the Picker buttons was pressed
+		if (this.pickerButtonPressed) {
+			forwardMouseUp(ev, this.props);
+			this.pickerButtonPressed = false;
 		}
 	}
 
 	handleDecDown = (ev) => {
 		if (ev) {
 			forwardMouseDown(ev, this.props);
+			this.pickerButtonPressed = true;
 		}
-		this.handleDown(-1);
+		this.handleDecrement();
 	}
 
 	handleIncDown = (ev) => {
 		if (ev) {
 			forwardMouseDown(ev, this.props);
+			this.pickerButtonPressed = true;
 		}
-		this.handleDown(1);
+		this.handleIncrement();
 	}
 
 	handleWheel = (ev) => {
@@ -555,13 +548,13 @@ const PickerBase = class extends React.Component {
 		}
 	}
 
-	throttleInc = new Job(this.handleIncClick, 200)
+	throttleInc = new Job(this.handleIncrement, 200)
 
-	throttleDec = new Job(this.handleDecClick, 200)
+	throttleDec = new Job(this.handleDecrement, 200)
 
-	throttleWheelInc = new Job(this.handleIncClick, 100)
+	throttleWheelInc = new Job(this.handleIncrement, 100)
 
-	throttleWheelDec = new Job(this.handleDecClick, 100)
+	throttleWheelDec = new Job(this.handleDecrement, 100)
 
 	handleKeyDown = (ev) => {
 		const {
@@ -807,6 +800,7 @@ const PickerBase = class extends React.Component {
 				onFocus={this.handleFocus}
 				onKeyDown={this.handleKeyDown}
 				onKeyUp={this.handleKeyUp}
+				onMouseUp={this.handleUp}
 				ref={this.initContainerRef}
 			>
 				<PickerButton
@@ -817,11 +811,9 @@ const PickerBase = class extends React.Component {
 					hidden={reachedEnd}
 					icon={incrementIcon}
 					joined={joined}
-					onClick={this.handleIncClick}
 					onHoldPulse={this.handleIncPulse}
 					onKeyDown={this.handleIncKeyDown}
 					onMouseDown={this.handleIncDown}
-					onMouseUp={this.handleUp}
 					onSpotlightDisappear={onIncrementSpotlightDisappear}
 					spotlightDisabled={spotlightDisabled}
 				/>
@@ -853,11 +845,9 @@ const PickerBase = class extends React.Component {
 					hidden={reachedStart}
 					icon={decrementIcon}
 					joined={joined}
-					onClick={this.handleDecClick}
 					onHoldPulse={this.handleDecPulse}
 					onKeyDown={this.handleDecKeyDown}
 					onMouseDown={this.handleDecDown}
-					onMouseUp={this.handleUp}
 					onSpotlightDisappear={onDecrementSpotlightDisappear}
 					spotlightDisabled={spotlightDisabled}
 				/>