+++ packages/moonstone/internal/Picker/tests/Picker-specs.js
@@ -18,13 +18,13 @@ describe('Picker Specs', function () {
 		expect(actual).to.equal(expected);
 	});
 
-	it('should return an object {value: Number} that represents the next value of the Picker component when clicking the increment <span>', function () {
+	it('should return an object {value: Number} that represents the next value of the Picker component when pressing the increment <span>', function () {
 		const handleChange = sinon.spy();
 		const picker = mount(
 			<Picker onChange={handleChange} min={-1} max={1} value={0} index={0} />
 		);
 
-		picker.find(`.${css.incrementer}`).simulate('click');
+		picker.find(`.${css.incrementer}`).simulate('mouseDown');
 
 		const expected = 1;
 		const actual = handleChange.args[0][0].value;
@@ -32,13 +32,13 @@ describe('Picker Specs', function () {
 		expect(actual).to.equal(expected);
 	});
 
-	it('should return an object {value: Number} that represents the next value of the Picker component when clicking the decrement <span>', function () {
+	it('should return an object {value: Number} that represents the next value of the Picker component when pressing the decrement <span>', function () {
 		const handleChange = sinon.spy();
 		const picker = mount(
 			<Picker onChange={handleChange} min={-1} max={1} value={0} index={0} />
 		);
 
-		picker.find(`.${css.decrementer}`).simulate('click');
+		picker.find(`.${css.decrementer}`).simulate('mouseDown');
 
 		const expected = -1;
 		const actual = handleChange.args[0][0].value;
@@ -52,7 +52,7 @@ describe('Picker Specs', function () {
 			<Picker onChange={handleChange} disabled min={0} max={0} value={0} index={0} />
 		);
 
-		picker.find(`.${css.incrementer}`).simulate('click');
+		picker.find(`.${css.incrementer}`).simulate('mouseDown');
 
 		const expected = false;
 		const actual = handleChange.called;
@@ -66,7 +66,7 @@ describe('Picker Specs', function () {
 			<Picker onChange={handleChange} wrap min={-1} max={0} value={0} index={0} />
 		);
 
-		picker.find(`.${css.incrementer}`).simulate('click');
+		picker.find(`.${css.incrementer}`).simulate('mouseDown');
 
 		const expected = -1;
 		const actual = handleChange.args[0][0].value;
@@ -80,7 +80,7 @@ describe('Picker Specs', function () {
 			<Picker onChange={handleChange} wrap min={0} max={1} value={0} index={0} />
 		);
 
-		picker.find(`.${css.decrementer}`).simulate('click');
+		picker.find(`.${css.decrementer}`).simulate('mouseDown');
 
 		const expected = 1;
 		const actual = handleChange.args[0][0].value;
@@ -96,7 +96,7 @@ describe('Picker Specs', function () {
 		const button = picker.find(`.${css.incrementer}`);
 
 		const expected = 3;
-		button.simulate('click');
+		button.simulate('mouseDown');
 		const actual = handleChange.args[0][0].value;
 
 		expect(actual).to.equal(expected);
@@ -110,7 +110,7 @@ describe('Picker Specs', function () {
 		const button = picker.find(`.${css.decrementer}`);
 
 		const expected = 0;
-		button.simulate('click');
+		button.simulate('mouseDown');
 		const actual = handleChange.args[0][0].value;
 
 		expect(actual).to.equal(expected);
@@ -122,7 +122,7 @@ describe('Picker Specs', function () {
 			<Picker onChange={handleChange} wrap step={3} min={0} max={3} value={3} index={0} />
 		);
 
-		picker.find(`.${css.incrementer}`).simulate('click');
+		picker.find(`.${css.incrementer}`).simulate('mouseDown');
 
 		const expected = 0;
 		const actual = handleChange.args[0][0].value;
@@ -136,7 +136,7 @@ describe('Picker Specs', function () {
 			<Picker onChange={handleChange} wrap step={3} min={0} max={9} value={0} index={0} />
 		);
 
-		picker.find(`.${css.decrementer}`).simulate('click');
+		picker.find(`.${css.decrementer}`).simulate('mouseDown');
 
 		const expected = 9;
 		const actual = handleChange.args[0][0].value;