+++ packages/moonstone/CHANGELOG.md
@@ -6,7 +6,12 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Added
 
+- `moonstone/Button` property value to `backgroundOpacity` called "lightTranslucent" to better serve colorful image backgrounds behind Buttons. This also affects `moonstone/IconButton` and `moonstone/Panels/ApplicationCloseButton`.
+- `moonstone/Panels` property `closeButtonBackgroundOpacity` to support `moonstone/Panels/ApplicationCloseButton`'s `backgroundOpacity` prop
+
 ### Changed
+
+- `moonstone/Panels/ApplicationCloseButton` to expose its `backgroundOpacity` prop
 
 ### Fixed